Reorganize directory structure; make tests work
Now it can be run and tested locally without installing.
