+++ lettersmith/hash.lua
