+++ tests/table_utils.lua
@@ -3,4 +3,4 @@ local suite = microtest.suite
 local test = microtest.test
 local equal = microtest.equal
 
-local table_utils = require("table_utils")
+local table_utils = require("lettersmith.table_utils")
